server/pyproject.toml
[project]
name = "agent-control-server"
version = "3.0.0"
description = "Server for Agent Control - manage and evaluate controls for AI agents"
requires-python = ">=3.12"
# Note: agent-control-models, agent-control-engine, and agent-control-evaluators are bundled at build time
dependencies = [
    "fastapi>=0.109.0",
    "starlette-exporter>=0.23.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "SQLAlchemy>=2.0.0",
    "psycopg[binary]>=3.1",
    "asyncpg>=0.29.0",
    "greenlet>=3.0",
    "alembic>=1.13.0",
    "jsonschema>=4.25.1",
    "jsonschema-rs>=0.22.0",
    "google-re2>=1.1",  # For engine (bundled)
    "sqlglot[rs]>=20.0.0",  # For SQL evaluator (bundled)
]
authors = [
    {name = "Agent Control Team"}
]
readme = "README.md"
license = {text = "Apache-2.0"}

[project.optional-dependencies]
luna2 = [
    "httpx>=0.24.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-jsonschema>=4.23.0",
    "agent-control-models",
    "agent-control-engine",
    "agent-control-evaluators",
]

[project.scripts]
agent-control-server = "agent_control_server.main:run"



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Note: agent_control_models, agent_control_engine, and agent_control_evaluators are copied by scripts/build.py
packages = ["src/agent_control_server", "src/agent_control_models", "src/agent_control_engine", "src/agent_control_evaluators"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["agent_control_server"]

# For local development, use workspace packages
[tool.uv.sources]
agent-control-models = { workspace = true }
agent-control-engine = { workspace = true }
agent-control-evaluators = { workspace = true }
